refactor: 컨트롤러 메서드 추가
