+++ app/src/main/java/com/techcourse/DispatcherServletInitializer.java
@@ -4,6 +4,7 @@ import jakarta.servlet.ServletContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import web.org.springframework.web.WebApplicationInitializer;
+import webmvc.org.springframework.web.servlet.DispatcherServlet;
 
 /**
  * Base class for {@link WebApplicationInitializer}
@@ -25,9 +26,9 @@ public class DispatcherServletInitializer implements WebApplicationInitializer {
                     "Check if there is another servlet registered under the same name.");
         }
 
-        registration.setLoadOnStartup(1);
         registration.addMapping("/");
 
         log.info("Start AppWebApplication Initializer");
     }
 }
+
