+++ app/src/main/java/com/techcourse/repository/InMemoryUserRepository.java
@@ -15,13 +15,17 @@ public class InMemoryUserRepository {
         database.put(user.getAccount(), user);
     }
 
+    private InMemoryUserRepository() {
+    }
+
     public static void save(User user) {
         database.put(user.getAccount(), user);
     }
 
     public static Optional<User> findByAccount(String account) {
+        if (account == null) {
+            return Optional.empty();
+        }
         return Optional.ofNullable(database.get(account));
     }
-
-    private InMemoryUserRepository() {}
 }
